pyproject.toml
[project]
name = "upgradeupstox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
     "asyncpg>=0.31.0",
    "dash>=3.3.0",
    "dash-extensions>=2.0.4",
    "fastapi[standard]>=0.122.0",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "protobuf>=6.33.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "sqlalchemy>=2.0.44",
    "upstox-python-sdk>=2.19.0",
    "websockets>=15.0.1",
]
